+++ CHANGELOG.md
@@ -24,6 +24,7 @@ Released on Sunday, July 26 2026.
 - Added support for `padding-block` and `padding-inline` declarations
 - Added missing box model declarations (e.g., `aspect-ratio`, `block-size`, `overflow-clip-margin`)
 - Added missing logical declarations (e.g., `inset`, `border-inline`, `border-end-start-radius`)
+- Added missing alignment declarations (e.g., `justify-items`, `place-content`)
 - Added more CSSOM possibilities and helpers (#6)
 - Added parts of recent color spec update such as `rgb` with spaces (#131)
 - Added now Color L4 parsing with `hsl`, `hwb`, `lab`, `lch`, `oklab`, and `oklch`